requirements.txt
# Core dependencies
# huggingface_hub range matches transformers 4.57.x and whisperx (hub 1.x breaks those pins).
huggingface_hub>=0.34.0,<1.0
torch>=2.0.0
numpy>=1.24.0

# Qwen3-TTS stack: PyPI qwen-tts pins these; listing them here prevents another package
# from leaving a stale transformers (ImportError: auto_docstring; check_model_inputs mismatch).
transformers==4.57.3
accelerate==1.12.0

# Qwen3-TTS (when TTS_BACKEND=qwen3)
qwen-tts>=0.1.1
soundfile>=0.12.0

# Audio quality analysis (background music, noise, recording quality)
librosa>=0.10.0

# API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
pydub>=0.25.1
# Python 3.13+ compatibility: stdlib `audioop` removed (pydub fallback)
audioop-lts>=0.2.1; python_version >= "3.13"
ffmpeg-python>=0.2.0
pedalboard>=0.9
pyloudnorm>=0.1
# System dependency required for audio conversion/mixing:
# - macOS: brew install ffmpeg
# - Ubuntu/Debian: apt install ffmpeg
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# Article scraping and LLM dependencies
beautifulsoup4>=4.12.0
httpx>=0.25.0
lxml>=4.9.0
# Audio transcription (used for style-oriented voice profiling from audio)
# Note: onnxruntime is not available for some very new Python versions yet.
# We keep transcription optional on Python >= 3.13 (the API will still work,
# but will return a minimal profile if transcription isn't available).
faster-whisper>=1.1.1; python_version < "3.13"
# Optional: newspaper3k for fallback scraping (requires lxml_html_clean)
# Uncomment the following lines if you want to use newspaper3k as a fallback scraper:
# newspaper3k>=0.2.8
# lxml_html_clean>=0.1.0

# Optional: FlashAttention2 for faster inference (recommended but not required)
# Install with: pip install flash-attn --no-build-isolation
# Note: Requires CUDA and may need specific CUDA/PyTorch versions
# If not installed, the model will fall back to SDPA (scaled dot product attention)

# Legacy VibeVoice (when TTS_BACKEND=vibevoice): clone VibeVoice repo and run pip install -e VibeVoice/
# Optional: flash-attn for lower GPU memory with Qwen3-TTS (pip install flash-attn --no-build-isolation)

# Transcript service dependencies
whisperx>=3.1.0
pyannote.audio>=3.1.0
pyannote.core>=5.0.0
speechbrain>=1.0.0
anthropic>=0.25.0
reportlab>=4.0.0
